Move classes related to room in more appropriate locations
