+++ app/src/androidTest/java/ch/epfl/sweng/eventmanager/AppDatabaseTests.java
@@ -8,9 +8,9 @@ import android.content.Context;
 import android.support.annotation.Nullable;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
-import ch.epfl.sweng.eventmanager.room.AppDataBase;
-import ch.epfl.sweng.eventmanager.room.data.JoinedEvent;
-import ch.epfl.sweng.eventmanager.room.data.JoinedEventDao;
+import ch.epfl.sweng.eventmanager.repository.room.AppDataBase;
+import ch.epfl.sweng.eventmanager.repository.data.JoinedEvent;
+import ch.epfl.sweng.eventmanager.repository.room.JoinedEventDao;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
