+++ app/src/main/java/ch/epfl/sweng/eventmanager/inject/ApplicationComponent.java
@@ -2,7 +2,7 @@ package ch.epfl.sweng.eventmanager.inject;
 
 import android.app.Application;
 import ch.epfl.sweng.eventmanager.EventManagerApplication;
-import ch.epfl.sweng.eventmanager.room.RoomModule;
+import ch.epfl.sweng.eventmanager.repository.room.RoomModule;
 import dagger.BindsInstance;
 import dagger.Component;
 import dagger.android.AndroidInjectionModule;
